Feature/feature active server monitor (#57)
* Added server monitor functionality

* Improved code and added model functionality

Added variables to models so they can be used with XML tags and added switchcases which can be used with ServerMonitorType. Depending on the type choosen it will run that specific Server Monitor.

* Check for threshold

Added a basic check to see if the diskThreshold has already been reached before. So it wont constantly resend the email

* Added email for disk check

Managed to get email to work so now when your below the threshold for disk space it will send an email to given email.

* Added check to see if disk has been sent and RAM check.

Added a dynamic boolean check for disk which checks if for that specific disk a email already has been sent. also added a simple check for the RAM space and how much is left. This will also now send an email to the user informing them about the RAM.

* Cleaned up code and added cpu check

Cleaned up the code and placed every part of the server monitor in its own function also added the cpu check and 2 different ways to check for it which can be chosen in the XML file

* Added the option for either all hard drives or a unique specified one.

I changed the Hard drive function so now the user can either give a specific drive name like C: or say they want all the drives that are installed.

* Added network interface check.

Added the network interface check which checks for the network utilization and added another Server monitor type so users can choose Network.

* Code cleanup.

Cleaned up some of the code and removed some unnecessary stuff.

* Added some comments for more info.

* Added function descriptions

* Improved code based on PR feedback

Made the code more inline with the programming guidlines.
Also made some changes to the body's send by the emails

* More changes based on PR feedback.

* Added JObjects and made CpuCounter adjustable

I added JObjects to every type of hardware check now. So now the results can be used elsewhere. The CpuCountArrray size can also be adjusted now via XML

* Made CPU counter value adjustable and small fixes

Now the CPU counter value can be adjusted in the XML. Also changed a few more variables to me var's. Also updated the ICommunicationService.

* Small whitespace fix

* Small fixes/improvements

fixed some spelling mistakes and removes some variables that werent used.

* Some code cleanup

---------

Co-authored-by: Gilian <giliankeulens@happyhorizon.com>