Feature/auto updater improvements (#177)
* Move successful update log outside of mail check

If the WTS was updated it would only be logged if a mail is being send, making it look like the update process was still busy.

* Add machine name to email subject to distinguish WTS instances

If the WTS instance has the default name or runs on multiple servers with the same name there was no indication in the email which was succeeded or failed. The machine name will solve this, since it indicates exactly which server the WTS is on.

* If the WTS exe file is missing always update it

If for some reason the exe file is not in the given directory, consider it to need to be updated. Since there was no exe file the auto updater won't try to start it.

* Prevent updates if the download failed